+++ Linux/Client/inc/Cache.h
@@ -2,7 +2,7 @@
 #define CACHE
 
 FILE* createTempFile(char* basename);
-int readRequest(MyDFSId* id, int pos, int size, CacheRequest* req);
+int readRequest(MyDFSId* id, long pos, unsigned int size, CacheRequest* req);
 int writeCache(MyDFSId* id, void* buffer, int size, int pos);
 
 #endif
